Move `RewardsCommand` to its own package
